backend/requirements.txt
asgiref==3.5.2
chardet>=3.0.0
Django==4.1.2
djangorestframework==3.14.0
multimetric==1.3.0
Pygments>=2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
pylama>=8.4.1
